Add public inits and allow mutating diff elements
On some scenarios, it's useful to be able to create or mutate a `Diff`
instance (or another variant), for example:
  - exclude some changes from the calculated diff
  - manually build our own diff, for simple scenarios

Since `init(elements:)` is auto generated by the compiler, it gets the
default `internal` visibility, meaning it's not accessible outside
`Differ`'s module. Adding a `public init(elements:)` enables this usage.

Additional `ExpressibleByArrayLiteral` conformance was added, to make
the creation of custom instances even easier.

The `elements` property was made `var`, because it enables mutating the
`struct` when the parent instance is itself mutable.

The following changes was applied only to `Diff`, `NestedDiff` and
`NestedExtendedDiff`. `ExtendedDiff` was not modified because it
contains additional properties (i.e. `source`, `sourceIndex`,
`reorderedIndex` and `moveIndices`) which are in `ExtendedPatch`
operations.