Differentiate platform targets
When integrating Differ by downloading the repo or as a git submodule, the current state of the project only contained a Mac target named Differ. This made it impossible to integrate with a non-macOS app target without making changes to Differ first. This change duplicates the existing Mac target in order to create iOS, tvOS and watchOS targets. The target names include the platform they are for, but the product name for all platforms is still Differ. This means that you can now download the repo or add Differ as a git submodule and start using it in a Mac, iOS, tvOS or watchOS project immediately.